requirements.txt
pytest==8.3.2
playwright==1.47.0
allure-pytest==2.13.5
python-dotenv==1.0.1
